requirements.txt
absl-py==1.4.0
accelerate==0.20.3
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
ansi2html==1.8.0
appdirs==1.4.4
APScheduler==3.10.1
arrow==1.2.3
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
audioread==3.0.0
av==9.2.0
Babel==2.12.1
backoff==1.11.1
beautifulsoup4==4.12.2
binaryornot==0.4.4
black==23.3.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
chex==0.1.6
chinese-converter==1.0.2
click==8.1.3
clldutils==3.19.0
cmaes==0.9.1
cmake==3.26.4
codecarbon==1.2.0
colorama==0.4.6
colorlog==6.7.0
commonmark==0.9.1
contourpy==1.1.0
cookiecutter==1.7.3
csvw==3.1.3
cycler==0.11.0
dash==2.11.0
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
datasets==2.13.1
decorator==5.1.1
decord==0.6.0
dill==0.3.4
dlinfo==1.2.1
dm-tree==0.1.8
evaluate==0.4.0
exceptiongroup==1.1.1
execnet==1.9.0
faiss-cpu==1.7.4
fastjsonschema==2.17.1
filelock==3.12.2
fire==0.5.0
Flask==2.2.5
flatbuffers==2.0.7
flax==0.5.3
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
fugashi==1.2.1
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.18
google-auth==2.21.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gql==3.4.1
graphql-core==3.2.3
greenlet==2.0.2
grpcio==1.51.3
h5py==3.9.0
hf-doc-builder==0.4.0
huggingface-hub==0.15.1
hypothesis==6.79.3
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
iniconfig==2.0.0
ipadic==1.0.0
isodate==0.6.1
isort==5.12.0
itsdangerous==2.1.2
jax==0.3.6
jaxlib==0.3.5
Jinja2==3.1.2
jinja2-time==0.2.0
joblib==1.2.0
jsonschema==4.17.3
jupyter_core==5.3.1
kenlm==0.1
keras==2.11.0
keras-nlp==0.5.2
kiwisolver==1.4.4
language-tags==1.2.0
lazy_loader==0.2
libclang==16.0.0
librosa==0.10.0.post2
lit==16.0.6
llvmlite==0.40.1
lxml==4.9.2
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.1
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.12.2
mypy-extensions==1.0.0
nbformat==5.9.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
numba==0.57.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
onnx==1.12.0
onnxconverter-common==1.13.0
opt-einsum==3.3.0
optax==0.1.4
optuna==3.2.0
packaging==23.1
pandas==2.0.2
parameterized==0.9.0
pathspec==0.11.1
phonemizer==3.2.1
Pillow==9.5.0
plac==1.3.5
platformdirs==3.8.0
plotly==5.15.0
pluggy==1.2.0
pooch==1.6.0
portalocker==2.0.0
poyo==0.5.0
protobuf==3.19.6
psutil==5.9.5
py-cpuinfo==9.0.0
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyctcdecode==0.5.0
Pygments==2.15.1
pygtrie==2.5.0
pylatexenc==2.10
pynvml==11.5.0
pyparsing==3.1.0
pypng==0.20220715.0
pyrsistent==0.19.3
pytest==7.4.0
pytest-timeout==2.1.0
pytest-xdist==3.3.1
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3
PyYAML==5.4.1
ray==2.5.1
rdflib==6.3.2
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
responses==0.18.0
retrying==1.3.4
rfc3986==1.5.0
rhoknp==1.3.0
rich==11.2.0
rjieba==0.1.11
rouge-score==0.1.2
rsa==4.9
ruff==0.0.259
sacrebleu==1.5.1
sacremoses==0.0.53
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.11.0
segments==2.2.1
sentencepiece==0.1.99
sigopt==8.8.0
six==1.16.0
smmap==5.0.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.4.1
soxr==0.3.5
SQLAlchemy==2.0.17
SudachiDict-core==20230110
SudachiPy==0.6.7
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
tensorflow==2.11.1
tensorflow-estimator==2.11.0
tensorflow-hub==0.13.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-text==2.11.0
tensorstore==0.1.39
termcolor==2.3.0
text-unidecode==1.3
tf2onnx==1.14.0
threadpoolctl==3.1.0
timeout-decorator==0.5.0
timm==0.9.2
tokenizers==0.13.3
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.65.0
traitlets==5.9.0
-e git+https://github.com/BorrisonXiao/transformers.git@7b4e3b5b40f20dfa9361460d5176447d448495b2#egg=transformers
triton==2.0.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==5.0.1
unidic==1.1.0
unidic-lite==1.0.8
uritemplate==4.1.1
urllib3==1.26.16
wasabi==0.10.1
Werkzeug==2.2.3
wrapt==1.15.0
xxhash==3.2.0
yarl==1.9.2
zipp==3.15.0
